+++ README.md
@@ -6,12 +6,14 @@
 
 Goland container to schedule and backup mysql database
 
-Example deploy on  ```deploy_sidecar_example/docker-compose.yml```
+Examples deploy on  ```examples/```
 
 Copy `env.sample` as `.env`
 
 ENVIROMENT VARIABLE   | DESCRIPTION | Values
 ----------   | ---------- | --------------  
+TARGET_FOLDER_PREFIX | folder and prefix filename pattern on upload S3 | `dumpfolder/prefix_` 
+SCHEDULE | see below | `0 * * * * *` once per minute
 AWS_ACCESS_KEY_ID | Aws access key or Minio username | Access key string
 AWS_SECRET_ACCESS_KEY | Aws secret key or Minio password | Secret access key string
 AWS_DEFAULT_REGION | Aws default region or any value for minio | `eu-west-1` etc
@@ -25,7 +27,6 @@ MYSQL_DATABASE | database name | string
 MYSQL_USER | database user | string
 MYSQL_PASSWORD | database password | string
 MYSQL_ALL_DB | cycle all database and backups single file each | `true` or `false`
-SCHEDULE | see below | 
 CLEAN_DAYS | number of backup retention days | integer or empty
 
 
@@ -44,7 +45,7 @@ Day of week  | Yes        | 0-6 or SUN-SAT  | * / , - ?
 
 
 
-# Usage
+# Usage, AWS Example
 
 Create `.env` file:
 
@@ -63,3 +64,28 @@ Launch with
 ```bash
 docker-compose up -d
 ```
+
+
+# Usage, MINIO Example
+
+Create `.env` file:
+
+```bash
+
+```
+
+Create `docker-compose.yml` file:
+
+```yml
+
+```
+
+Launch with
+
+```bash
+docker-compose up -d
+```
+
+# Kubernetes Chart repository
+
+Go to [beeckup/charts](https://github.com/beeckup/charts) for kubernetes chart
