--- deploy_sidecar_example/.env.sample
@@ -1,31 +1,0 @@
-### db connection
-MYSQL_HOST=db
-MYSQL_DATABASE=wordpress
-MYSQL_USER=root
-MYSQL_PASSWORD=123456
-MYSQL_SQL_FILENAME=nomebackup
-### cron schedule
-SCHEDULE=0 * * * * *
-
-### PUT S3_UPLOAD to true to upload your dump on s3 or minio bucket
-S3_UPLOAD=true
-### S3 or minio host
-S3_HOST=minio:9000
-### Protocol
-S3_PROTOCOL=http
-### Your bucket name
-S3_BUCKET=cicciopollo
-### minio or s3 credentials
-S3_KEY=85A8U57ZITLSLFBYKNCG
-S3_SECRET=14MAuAetrv7y3E6zAuUOimXy5KYRqrZKw3cWuEe/
-### port of local minio
-MINIO_PORT=9000
-
-### ZIP FILE
-ZIP_FILE=true
-
-### ALL DB , true to enable
-MYSQL_ALL_DB=
-
-### Number of days to maintain backup history
-CLEAN_DAYS=15
