+++ dev-run-to-aws.sh
@@ -14,4 +14,4 @@ SCHEDULE="*/10 * * * * *" \
   MYSQL_USER="root" \
   MYSQL_PASSWORD="123456" \
   MYSQL_ALL_DB="false" \
-  go run backup.go
+  go run src/main.go src/common.go src/minio.go src/aws.go src/mysql.go src/zip.go
