+++ docker/Dockerfile
@@ -1,0 +1,9 @@
+FROM golang:alpine
+WORKDIR /go/src/app
+COPY . .
+RUN mkdir tempdump
+RUN apk add --no-cache git mysql-client openssl curl python py-pip coreutils\
+    && go get github.com/robfig/cron \
+    && go get github.com/aws/aws-sdk-go
+RUN go build -o main src/main.go src/common.go src/minio.go src/aws.go src/mysql.go src/zip.go
+CMD ["./main"]
